separate shape and kinematics
