+++ include/r3d/r3d.h
@@ -39,6 +39,7 @@
 #include "r3d_model.h"
 #include "r3d_probe.h"
 #include "r3d_screen_shader.h"
+#include "r3d_shape.h"
 #include "r3d_skeleton.h"
 #include "r3d_sky.h"
 #include "r3d_surface_shader.h"
